+++ pkg/generator/generator.go
@@ -4,9 +4,9 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"html/template"
 	"os"
 	"path"
+	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
 	"github.com/ghodss/yaml"
@@ -381,7 +381,7 @@ func (g Generator) renderTemplate(ctx context.Context, templateText string, temp
 		return "", microerror.Mask(err)
 	}
 
-	funcMap := sprig.FuncMap()
+	funcMap := sprig.TxtFuncMap()
 	funcMap["include"] = g.include
 
 	t, err := template.New("main").Funcs(funcMap).Option("missingkey=error").Parse(templateText)
@@ -405,7 +405,7 @@ func (g Generator) include(templateName string, templateData interface{}) (strin
 		return "", microerror.Mask(err)
 	}
 
-	t, err := template.New(templateName).Funcs(sprig.FuncMap()).Option("missingkey=error").Parse(string(contents))
+	t, err := template.New(templateName).Funcs(sprig.TxtFuncMap()).Option("missingkey=error").Parse(string(contents))
 	if err != nil {
 		return "", microerror.Mask(err)
 	}