+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.10] - 2021-04-13
+
 ### Fixed
 
 - Use `text/template` instead of `html` to avoid escaping strings.
@@ -103,7 +105,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create CLI/daemon scaffolding.
 - Generate application configuration using `generate` command.
 
-[Unreleased]: https://github.com/giantswarm/config-controller/compare/v0.2.9...HEAD
+[Unreleased]: https://github.com/giantswarm/config-controller/compare/v0.2.10...HEAD
+[0.2.10]: https://github.com/giantswarm/config-controller/compare/v0.2.9...v0.2.10
 [0.2.9]: https://github.com/giantswarm/config-controller/compare/v0.2.8...v0.2.9
 [0.2.8]: https://github.com/giantswarm/config-controller/compare/v0.2.7...v0.2.8
 [0.2.7]: https://github.com/giantswarm/config-controller/compare/v0.2.6...v0.2.7